Clean cpp (#5)
* completed read from matrix.  This will allow for better integration with R

* binning is working with new input format.

* added bin parameter print and selection

* fixed issue when root node is pure

* finally able to grow forest from R input and predict from R input... currently segfaults on exit from R.

* restructured, added R functions to keep separate, need to test and change multicore imp for bins.

* added the additional two R files required

* delete R files

* delete experiments

* cleaning

+++ quickTest.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-rm error.log
+if [ -f error.log ]; then
+	rm error.log
+fi
 
-for i in `seq 1 5`;
+for i in `seq 1 8`;
 do
-		./bin/fp $i 1 1
-		if [ $? -ne 0 ]
-		then
-			echo $i $q $j >> error.log
-		fi
-		printf "\n"
+	./bin/fp $i 1 1 > garbage.out
+	if [ $? -ne 0 ]
+	then
+		echo $i >> error.log
+	fi
+	printf "\n"
 done
+
+rm garbage.out
+
+if [ -f error.log ]; then
+	echo "errors were found"	
+else
+	./res/showRes.sh
+fi
+
+
